refactor: Repository 추상화
