+++ app/src/main/java/woowacourse/shopping/data/CartRepositoryImpl.kt
@@ -1,19 +1,20 @@
 package woowacourse.shopping.data
 
 import woowacourse.shopping.model.Product
+import woowacourse.shopping.repository.CartRepository
 
 // TODO: Step2 - CartProductDao를 참조하도록 변경
-class CartRepositoryImpl(cartProduct: List<Product> = listOf()) {
+class CartRepositoryImpl(cartProduct: List<Product> = listOf()) : CartRepository {
     private val _cartProducts: MutableList<Product> = cartProduct.toMutableList()
-    fun addCartProduct(product: Product) {
+    override fun addCartProduct(product: Product) {
         _cartProducts.add(product)
     }
 
-    fun getAllCartProducts(): List<Product> {
+    override fun getAllCartProducts(): List<Product> {
         return _cartProducts.toList()
     }
 
-    fun deleteCartProduct(id: Int) {
+    override fun deleteCartProduct(id: Int) {
         _cartProducts.removeAt(id)
     }
 }
