+++ app/src/main/java/woowacourse/shopping/ui/MainViewModelFactory.kt
@@ -1,0 +1,18 @@
+package woowacourse.shopping.ui
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+import woowacourse.shopping.repository.CartRepository
+import woowacourse.shopping.repository.ProductRepository
+
+class MainViewModelFactory(
+    private val productRepository: ProductRepository,
+    private val cartRepository: CartRepository
+) : ViewModelProvider.Factory {
+    override fun <T : ViewModel?> create(modelClass: Class<T>): T {
+        if (modelClass.isAssignableFrom(MainViewModel::class.java)) {
+            return MainViewModel(productRepository, cartRepository) as T
+        }
+        throw IllegalArgumentException("Unknown ViewModel Class")
+    }
+}
